Made changes from pull request
1. moved to extras.the_wrong_way_cow package
2. findWrongWayCow returns null so it compiles.
3. added Junit to the classpath
